+++ lib/velocity_x.dart
@@ -24,6 +24,7 @@ export 'src/flutter/flex.dart';
 export 'src/flutter/gesture.dart';
 export 'src/flutter/icon.dart';
 export 'src/flutter/list.dart';
+export 'src/flutter/marquee.dart';
 export 'src/flutter/object.dart';
 export 'src/flutter/opacity.dart';
 export 'src/flutter/padding.dart';